requirements.txt
Beaker==1.6.4
BeautifulSoup==3.2.1
Chameleon==2.9.2
Flask==0.9
Markdown==2.1.1
MarkupSafe==0.15
Paste==1.7.5.1
PasteDeploy==1.5.0
PyJWT==0.1.4
Pycco==0.3.0
Pygments==1.5
SQLAlchemy==0.7.8
WebOb==1.2.2
Werkzeug==0.8.3
-e git+https://github.com/hypothesis/annotator-store.git@20c941b62e2a97b4cb0ab169b5d399adf636f3d2#egg=annotator-dev
anykeystore==0.2
argparse==1.2.1
certifi==0.0.8
chardet==1.0.1
clik==0.3.1
colander==0.9.8
cryptacular==1.4.1
deform==0.9.5
deform-bootstrap==0.2.2
hem==0.3.2
-e git+https://github.com/eventray/horus.git@0cd44871d20d10038e77080c894ea1a10298ecac#egg=horus-dev
httplib2==0.7.4
iso8601==0.1.4
mock==0.8.0
nose==1.1.2
oauth2==1.5.211
pbkdf2==1.3
peppercorn==0.4
py==1.4.9
pyes==0.19.1
pyramid==1.4
pyramid-basemodel==0.1.3
pyramid-beaker==0.7
pyramid-debugtoolbar==1.0.2
pyramid-deform==0.2a5
pyramid-layout==0.6
pyramid-mailer==0.9
pyramid-tm==0.4
-e git+https://github.com/sontek/pyramid_webassets.git@7ee74fa419fddc357c03d64a05b2e896dabad781#egg=pyramid_webassets-dev
pystache==0.5.2
pytest==2.2.4
python-dateutil==2.1
python-openid==2.2.5
repoze.lru==0.6
repoze.sendmail==3.2
requests==0.11.2
simplejson==2.6.2
smartypants==1.6.0.3
sockjs-tornado==1.0.0
transaction==1.3.0
translationstring==1.1
unittest2==0.5.1
velruse==1.0
venusian==1.0a7
webassets==0.8
wsgiref==0.1.2
zope.deprecation==4.0.0
zope.interface==4.0.1
zope.sqlalchemy==0.7.1
